Fix TypeScript bindings export for Specta compatibility
- Replace `Option<usize>` with `Option<u32>` in ProjectCommand::AddTrack
- Update both commands.rs and types_export.rs to use u32 instead of usize
- Fixes "BigIntForbidden" error during Tauri development startup
- Allows successful TypeScript bindings generation from Rust types

+++ src-tauri/src/types_export.rs
@@ -245,7 +245,7 @@ pub enum ProjectCommand {
   AddTrack {
     name: String,
     track_type: TrackType,
-    index: Option<usize>,
+    index: Option<u32>,
   },
   DeleteTrack {
     track_id: String,
